Move scootapi cmds
