+++ scheduler/client/kill_job.go
