+++ scheduler/client/offline_worker.go
